fix(amazonq): Prevent customization override if user has manually selected a customization
## Problem

If a user was part of an AB group with a customization override, they could only ever use the customization from the override regardless of personal choice. 

## Solution

Changes the logic for when a customization will be overridden if the user is part of an AB group that has a customization override. With this change, if the user has manually selected a customization, it will not be overridden by the AB customization.