feat(amazonq): capability to send new context commands to AB groups (#6239)
## Problem

When working on new features available to AB groups, the new action are
not easily discoverable to users

## Solution

Add the ability to send a new command that will appear when a user types
@ in the chat window

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).

License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.