deps: update "aws-sdk" to v2.1692.0 #6340
## Problem
The version of the AWSK SDK v2 was outdated 1.5 years, so some newer
functionality was not available.

## Solution
Update the version of `aws-sdk` package to the latest version.

The `Response` class in AWS SDK changed the signature of the `nextPage`
method, so some tests that were mocking this method had to be updated
accordingly (changing the response from `undefined` to `null`).

The definition changed from 
```
    nextPage(callback?: (err: E, data: D) => void): Request<D, E>|void;
```
(callback optional, can return `void`)

To 
```
    nextPage(): Request<D, E> | null;
    nextPage(callback: (err: E, data: D) => void): void;
```
(can return `null` if no callback. only when a callback is passed
returns `void`)