telemetry(auth): debounce `aws_loadCredentials` and add debugging information (#6499)
## Problem
This metric is currently being spammed in telemetry at unusual volume.
It is suspected that this is connected to this change:
https://github.com/aws/aws-toolkit-vscode/pull/5979/files, but the
connection is not yet clear.

## Solution
- use the `withTelemetryContext` decorator to emit a `function_call`
metric to capture additional information. This metric should allow us to
determine what is calling this function so much. Added techdebt test to
remove this debugging information before the following release.
- debounce the current emit line that is causing the high volume
emission.


## Future Work 
- `debounce` reimplements a special case of `debounceWithCancel`, these
can be redone to reduce code and improve reliability.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.