ci(lint): rule to detect index.ts imports in core
I could not find a lint rule (or a configuration of a lint rule) to achieve this, so I just made one. I tried some things from eslint-plugin-import (namely no-cycles), but I was not happy with the results.

This will prevent you from importing from index.ts files while working in packages/core/, which has been a source of annoying and hard to understand circular dependency issues.

It only affects imports in core, and only in src/ dirctories that do not start with test/