deps(cw): regen clients with codegen at 0.25.x (#6439)
The two streaming clients committed to the repo are generated via
different versions of `smithy-typescript-codegen`. Each version of
codegen pins the generated client to use specific versions of
`@aws-sdk/*` and `@smithy/*` packages. The root project also consumes
different versions of these packages. Therefore, the project is
currently consuming three different versions of many `@aws-sdk/*` and
`@smithy/*` packages.

This is problematic because it can cause dependency conflicts. Certain
versions of `@aws-sdk/*` and `@smithy/*` packages are only compatible
with certain other versions. For more information on this, see the
discussion involving help from an SDK team member explaining this
problem.

- Regenerate both clients with the same version of codegen `0.25.x`.
- Pin all necessary versions to ensure a single version of `@aws-sdk/*`
and `@smithy/*` packages.
- Add documentation to the build instructions that these clients must
stay in-sync (in terms of codegen).
- Add documentation regarding updating `@aws-sdk/*` packages.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).
- License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.